разделил загрузку и патч спеки в отдельные классы
